refactoring openfda for drugs
